+++ NEWS.md
@@ -1,3 +1,5 @@
+# zooper (development version)
+
 * Updated to newest SMSCG data
 
 # zooper 2.4.1