+++ NEWS.md
@@ -1,3 +1,8 @@
+# zooper (development version)
+
+* Add option to specify download method in `Zoopdownloader` function.
+* Updated to newest SMSCG data
+
 # zooper 2.4.1
 
 * Updated to newest FMWT macro column names.