+++ NEWS.md
@@ -1,3 +1,14 @@
+# zooper (development version)
+
+* Updating to newest EMP data. 
+* Added newest FMWT/STN data, including new station STN 722
+* Add option to specify download method in `Zoopdownloader` function.
+* Updated to newest SMSCG data
+
+# zooper 2.4.1
+
+* Updated to newest FMWT macro column names.
+
 # zooper 2.4.0
 
 * Changed all *Acanthocyclops vernalis* in crosswalk to *Acanthocyclops*.